# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "golang作品",
        "title": "Gin-Backend",
        "termsOfService": "github",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/product/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Order AddProduct",
                "parameters": [
                    {
                        "description": "Add Product",
                        "name": "p",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/product/delete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Order DeleteProduct",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete Product",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/product/edit": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Order EditProduct",
                "parameters": [
                    {
                        "description": "Edit Product",
                        "name": "p",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/product/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Order ProductInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product Info",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/product/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Order ProductList",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "內部錯誤",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Product": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "categoryName": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "originalPrice": {
                    "type": "number"
                },
                "productBanner": {
                    "type": "string"
                },
                "productCoverImg": {
                    "type": "string"
                },
                "productDetailContent": {
                    "type": "string"
                },
                "productId": {
                    "type": "string"
                },
                "productIntro": {
                    "type": "string"
                },
                "productName": {
                    "type": "string"
                },
                "sellStatus": {
                    "type": "integer"
                },
                "sellingPrice": {
                    "type": "number"
                },
                "stockNum": {
                    "type": "integer"
                },
                "tag": {
                    "type": "string"
                },
                "updateUser": {
                    "type": "string"
                }
            }
        },
...
    }
}